Correccion de Errores, Listas de Tipo de Animales
