+++ Mascotapp/Mascotapp/Mascotapp.csproj
@@ -49,6 +49,9 @@
     <EmbeddedResource Update="Solicitudes\VisualizarSolicitud.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
+    <EmbeddedResource Update="Tipo animales\TipoDeAnimal.xaml">
+      <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
+    </EmbeddedResource>
     <EmbeddedResource Update="Visualizar mapa\Mapa.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
