+++ Mascotapp/Mascotapp/Tipo animales/NuevoTipo.xaml
@@ -4,33 +4,24 @@
              xmlns:d="http://xamarin.com/schemas/2014/forms/design"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              mc:Ignorable="d"
-             x:Class="Mascotapp.Tipo_animales.TipoAnimales">
+             x:Class="Mascotapp.Tipo_animales.TipoAnimal">
     <ContentPage.Content>
         <StackLayout Orientation="Vertical">
-            <Label Text="Tipo de animal" HorizontalOptions="CenterAndExpand" />
+            <Label Text="Nuevo Tipo de animal" HorizontalOptions="CenterAndExpand" />
 
             <StackLayout BackgroundColor="GhostWhite">
 
-                <StackLayout Orientation="Horizontal">
-                    <Label Text="Tipo animal: "/>
-                    <Picker  Title="Seleccione tipo animal" >
-                        <Picker.Items>
-                            <x:String>Gato1</x:String>
-                            <x:String>Gatoo2</x:String>
-                        </Picker.Items>
-                    </Picker>
-                </StackLayout>
-
                 <StackLayout Orientation="Vertical">
                     <Label Text="Descripcion: "/>
-                    <Editor Placeholder="Descripcion"></Editor>
+                    <Editor x:Name="txtNuevaDescripcion"
+                            Placeholder="Descripcion"/>
                 </StackLayout>
 
                 <StackLayout Orientation="Horizontal">
-                    <Button Text="Guardar"></Button>
-                    <Button Text="Cancelar"></Button>
+                    <Button x:Name="btnGuardar" Text="Guardar"></Button>
+                    <Button x:Name="btnCancelar" Text="Cancelar"></Button>
                 </StackLayout>
-
+                
             </StackLayout>
         </StackLayout>
     </ContentPage.Content>
