+++ Mascotapp/Mascotapp/Tipo animales/NuevoTipo.xaml.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using Domain.Entidades;
+using Domain.Servicios;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -10,11 +12,33 @@ using Xamarin.Forms.Xaml;
 namespace Mascotapp.Tipo_animales
 {
     [XamlCompilation(XamlCompilationOptions.Compile)]
-    public partial class TipoAnimales : ContentPage
+    public partial class TipoAnimal : ContentPage
     {
-        public TipoAnimales()
+        private ServicioTipoAnimal serviceTipoAnimal = new ServicioTipoAnimal();
+
+        public TipoAnimal()
         {
             InitializeComponent();
+            CargarEvento();
+        }
+
+        void CargarEvento()
+        {
+            btnGuardar.Clicked += btnGuardar_Clicked;
+            btnCancelar.Clicked += btnCancelar_Clicked;
+        }
+
+        private void btnGuardar_Clicked(object sender, EventArgs e)
+        {
+            var tipoAnimal = new Domain.Entidades.TipoAnimal();
+            tipoAnimal.IdTipoAnimal = 1;//pckTipoAnimal.SelectedItem.
+            tipoAnimal.Descripcion = txtNuevaDescripcion.Text;
+            serviceTipoAnimal.GuardarTipoAnimal(tipoAnimal);
+        }
+
+        private void btnCancelar_Clicked(object sender, EventArgs e)
+        {
+            throw new NotImplementedException();
         }
     }
 }
