+++ TipoAnimal.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Domain.Entidades
+{
+    public class TipoAnimal
+    {
+        public int? IdTipoAnimal { get; set; }
+        public string Descripcion { get; set; }
+    }
+}
